cormen/src/test/java/td/problemsolving/interviewbit/LongestCommonPrefix/SolutionTest.java
package td.problemsolving.interviewbit.LongestCommonPrefix;

import com.google.common.collect.ImmutableList;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;

import static org.junit.jupiter.api.Assertions.*;

class SolutionTest {
    @Test
    void longestCommonPrefix() {
        ArrayList<String> input;

        input = new ArrayList<>(ImmutableList.of("aaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"));
        assertEquals("aaaaaaa", new Solution().longestCommonPrefix(input));

        input = new ArrayList<>(ImmutableList.of("abcdefgh", "abcefgh"));
        assertEquals("abc", new Solution().longestCommonPrefix(input));

        input = new ArrayList<>(ImmutableList.of("abcdefgh", "aefghijk", "abcefgh"));
        assertEquals("a", new Solution().longestCommonPrefix(input));

        input = new ArrayList<>(ImmutableList.of("abab", "ab", "abcd"));
        assertEquals("ab", new Solution().longestCommonPrefix(input));
    }
}
